PROJECT_NAME/requirements.txt
boto
django
django-compressor
django-libsass
django-storages
djangorestframework
gunicorn
psycopg2
raven
